+++ base/linalg.jl
@@ -222,10 +222,6 @@ include("linalg/special.jl")
 include("linalg/bitarray.jl")
 include("linalg/ldlt.jl")
 
-include("linalg/sparse.jl")
-include("linalg/umfpack.jl")
-include("linalg/cholmod.jl")
-
 include("linalg/arpack.jl")
 include("linalg/arnoldi.jl")
