+++ base/sparse.jl
@@ -3,8 +3,11 @@ include("sparse/abstractsparse.jl")
 module SparseMatrix
 
 importall Base
+importall Base.LinAlg
+
 import Base.NonTupleType, Base.float, Base.Order, Base.Sort.Forward
 import Base.transpose!, Base.ctranspose!
+import Base.LinAlg.TriangularUnion
 
 export SparseMatrixCSC,
        blkdiag, dense, diag, diagm, droptol!, dropzeros!, etree, full,
@@ -15,5 +18,9 @@ export SparseMatrixCSC,
 
 include("sparse/sparsematrix.jl")
 include("sparse/csparse.jl")
+
+include("sparse/linalg.jl")
+include("sparse/umfpack.jl")
+include("sparse/cholmod.jl")
 
 end # module SparseMatrix
